Move worksearch.search.work_wrapper to .subjects
Only used there
