+++ openlibrary/plugins/worksearch/subjects.py
@@ -10,6 +10,7 @@ from infogami.utils.view import render_template, safeint
 
 from openlibrary.core.models import Subject
 from openlibrary.core.lending import add_availability
+from openlibrary.solr.query_utils import query_dict_to_str
 from openlibrary.utils import str_to_key, finddict
 
 
@@ -241,8 +242,6 @@ class SubjectEngine:
     ):
         # Circular imports are everywhere -_-
         from openlibrary.plugins.worksearch.code import run_solr_query
-        from openlibrary.plugins.worksearch.search import work_wrapper
-        from openlibrary.solr.query_utils import query_dict_to_str
 
         meta = self.get_meta(key)
         subject_type = meta.name
@@ -309,7 +308,7 @@ class SubjectEngine:
             name=name,
             subject_type=subject_type,
             work_count=result.num_found,
-            works=add_availability([work_wrapper(d) for d in result.docs]),
+            works=add_availability([self.work_wrapper(d) for d in result.docs]),
         )
 
         if details:
@@ -395,6 +394,36 @@ class SubjectEngine:
             return [value, count]
         else:
             return web.storage(name=value, count=count)
+
+    @staticmethod
+    def work_wrapper(w: dict) -> web.storage:
+        """
+        Convert a solr document into the doc returned by the /subjects APIs.
+        These docs are weird :/ We should be using more standardized results
+        across our search APIs, but that would be a big breaking change.
+        """
+        ia_collection = w.get('ia_collection_s', '').split(';')
+        return web.storage(
+            key=w['key'],
+            title=w["title"],
+            edition_count=w["edition_count"],
+            cover_id=w.get('cover_i'),
+            cover_edition_key=w.get('cover_edition_key'),
+            subject=w.get('subject', []),
+            ia_collection=ia_collection,
+            lendinglibrary='lendinglibrary' in ia_collection,
+            printdisabled='printdisabled' in ia_collection,
+            lending_edition=w.get('lending_edition_s', ''),
+            lending_identifier=w.get('lending_identifier_s', ''),
+            authors=[
+                web.storage(key=f'/authors/{olid}', name=name)
+                for olid, name in zip(w.get('author_key', []), w.get('author_name', []))
+            ],
+            first_publish_year=w.get('first_publish_year'),
+            ia=w.get('ia', [None])[0],
+            public_scan=w.get('public_scan_b', bool(w.get('ia'))),
+            has_fulltext=w.get('has_fulltext', False),
+        )
 
 
 def setup():
